# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
  - name: pairing
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /pairing/create:
    post:
      operationId: createPairing
      tags: [pairing]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/PairingCreateInput" }
      responses:
        "200":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Session" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "429": { $ref: "#/components/responses/RateLimited" }
  /pairing/join:
    post:
      operationId: joinPairing
      tags: [pairing]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/PairingJoinInput" }
      responses:
        "200":
          description: Joined
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Session" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409":
          description: Legacy room cannot accept new members
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "410":
          description: Room has ended
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "429": { $ref: "#/components/responses/RateLimited" }
  /pairing/session:
    get:
      operationId: getPairingSession
      tags: [pairing]
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: Current session
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Session" }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /pairing/assignment:
    patch:
      operationId: updatePairingAssignment
      tags: [pairing]
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/PairingAssignmentInput" }
      responses:
        "200":
          description: Updated current member assignment
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Session" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /pairing/command:
    post:
      operationId: sendPairingCommand
      tags: [pairing]
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/PairingCommandInput" }
      responses:
        "200":
          description: Updated or idempotently replayed session
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Session" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
  /pairing/rotate-code:
    post:
      operationId: rotatePairingCode
      tags: [pairing]
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: Rotated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Session" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
  /pairing/leave:
    post:
      operationId: leavePairing
      tags: [pairing]
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: Revoked
          content:
            application/json:
              schema: { $ref: "#/components/schemas/OkResponse" }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /pairing/end:
    post:
      operationId: endPairing
      tags: [pairing]
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: Room ended and all memberships revoked
          content:
            application/json:
              schema: { $ref: "#/components/schemas/OkResponse" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque
  responses:
    BadRequest:
      description: Malformed input
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
    Unauthorized:
      description: Invalid authentication
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
    Forbidden:
      description: Forbidden
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
    NotFound:
      description: Not found
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
    RateLimited:
      description: Too many requests
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
  schemas:
    HealthStatus:
      type: object
      additionalProperties: false
      required: [status]
      properties:
        status: { type: string }
    ErrorResponse:
      type: object
      additionalProperties: false
      required: [error]
      properties:
        error: { type: string }
    OkResponse:
      type: object
      additionalProperties: false
      required: [ok]
      properties:
        ok: { type: boolean }
    Group:
      type: object
      required: [id, name, score]
      properties:
        id: { type: string, minLength: 1, maxLength: 80 }
        name: { type: string, minLength: 1, maxLength: 100 }
        score: { type: number, minimum: 0, maximum: 1000000000 }
        color: { type: string, maxLength: 200 }
        badgeColor: { type: string, maxLength: 200 }
        icon: { type: string, maxLength: 30 }
    HistoryEntry:
      type: object
      required: [id, groupId, groupName, amount, reason, timestamp]
      properties:
        id: { type: string, minLength: 1, maxLength: 80 }
        groupId: { type: string, minLength: 1, maxLength: 80 }
        groupName: { type: string, minLength: 1, maxLength: 100 }
        amount: { type: number, minimum: -1000000, maximum: 1000000 }
        reason: { type: string, minLength: 1, maxLength: 500 }
        timestamp: { type: string, minLength: 1, maxLength: 100 }
    LapRecord:
      type: object
      required: [id, runnerName, group, minutes, seconds, ms, timeFormatted, totalSeconds, courseName, date, monthYear]
      properties:
        id: { type: string, minLength: 1, maxLength: 80 }
        runnerName: { type: string, minLength: 1, maxLength: 100 }
        group: { type: string, minLength: 1, maxLength: 100 }
        minutes: { type: integer, minimum: 0, maximum: 999 }
        seconds: { type: integer, minimum: 0, maximum: 59 }
        ms: { type: integer, minimum: 0, maximum: 99 }
        timeFormatted: { type: string, minLength: 1, maxLength: 30 }
        totalSeconds: { type: number, minimum: 0, maximum: 100000 }
        courseName: { type: string, minLength: 1, maxLength: 200 }
        date: { type: string, minLength: 1, maxLength: 30 }
        monthYear: { type: string, minLength: 1, maxLength: 100 }
    MonthlyRecord:
      type: object
      required: [id, month, winner, scores, rewardClaimed]
      properties:
        id: { type: string, minLength: 1, maxLength: 80 }
        month: { type: string, minLength: 1, maxLength: 100 }
        winner: { type: string, minLength: 1, maxLength: 500 }
        scores: { type: string, maxLength: 2000 }
        rewardClaimed: { type: boolean }
    Activity:
      type: object
      required: [id, title, type, points, location]
      properties:
        id: { type: string, minLength: 1, maxLength: 80 }
        title: { type: string, minLength: 1, maxLength: 300 }
        type: { type: string, minLength: 1, maxLength: 100 }
        points: { type: number, minimum: -1000000, maximum: 1000000 }
        location: { type: string, minLength: 1, maxLength: 100 }
        scrambledPhrase: { type: string, maxLength: 2000 }
        solvedPhrase: { type: string, maxLength: 2000 }
        hint: { type: string, maxLength: 2000 }
        lesson: { type: string, maxLength: 5000 }
        materials: { type: string, maxLength: 5000 }
        steps: { type: string, maxLength: 10000 }
        harder: { type: string, maxLength: 5000 }
        safety: { type: string, maxLength: 5000 }
    PairingState:
      type: object
      additionalProperties: false
      required: [groups, history, lapRecords, monthlyRecords, activities]
      properties:
        groups:
          type: array
          maxItems: 100
          items: { $ref: "#/components/schemas/Group" }
        history:
          type: array
          maxItems: 1000
          items: { $ref: "#/components/schemas/HistoryEntry" }
        lapRecords:
          type: array
          maxItems: 10000
          items: { $ref: "#/components/schemas/LapRecord" }
        monthlyRecords:
          type: array
          maxItems: 1000
          items: { $ref: "#/components/schemas/MonthlyRecord" }
        activities:
          type: array
          maxItems: 10000
          items: { $ref: "#/components/schemas/Activity" }
    Member:
      type: object
      additionalProperties: false
      required: [id, name, role, groupId, assignmentDate]
      properties:
        id: { type: string, format: uuid }
        name: { type: string, minLength: 1, maxLength: 60 }
        role: { type: string, enum: [owner, counselor] }
        groupId: { type: ["string", "null"], minLength: 1, maxLength: 80 }
        assignmentDate: { type: ["string", "null"], pattern: "^\\d{4}-(0[1-9]|1[0-2])-(0[1-9]|[12]\\d|3[01])$" }
    Session:
      type: object
      additionalProperties: false
      required: [roomId, code, role, memberId, groupId, assignmentDate, version, state, members]
      properties:
        roomId: { type: string, format: uuid }
        code: { type: string, pattern: "^[A-Za-z0-9]{10}$" }
        role: { type: string, enum: [owner, counselor] }
        memberId: { type: string, format: uuid }
        groupId: { type: ["string", "null"], minLength: 1, maxLength: 80 }
        assignmentDate: { type: ["string", "null"], pattern: "^\\d{4}-(0[1-9]|1[0-2])-(0[1-9]|[12]\\d|3[01])$" }
        version: { type: integer, minimum: 1 }
        state: { $ref: "#/components/schemas/PairingState" }
        members:
          type: array
          items: { $ref: "#/components/schemas/Member" }
        token: { type: string, minLength: 32 }
    PairingCreateInput:
      type: object
      additionalProperties: false
      required: [name, password, state, groupId, assignmentDate]
      properties:
        name: { type: string, minLength: 1, maxLength: 60 }
        password: { type: string, minLength: 4, maxLength: 128 }
        state: { $ref: "#/components/schemas/PairingState" }
        groupId: { type: ["string", "null"], minLength: 1, maxLength: 80 }
        assignmentDate: { type: string, pattern: "^\\d{4}-(0[1-9]|1[0-2])-(0[1-9]|[12]\\d|3[01])$" }
    PairingJoinInput:
      type: object
      additionalProperties: false
      required: [name, password, code, groupId, assignmentDate]
      properties:
        name: { type: string, minLength: 1, maxLength: 60 }
        password: { type: string, minLength: 4, maxLength: 128 }
        code: { type: string, pattern: "^[A-Za-z0-9]{10}$" }
        groupId: { type: string, minLength: 1, maxLength: 80 }
        assignmentDate: { type: string, pattern: "^\\d{4}-(0[1-9]|1[0-2])-(0[1-9]|[12]\\d|3[01])$" }
    PairingAssignmentInput:
      type: object
      additionalProperties: false
      required: [groupId, assignmentDate]
      properties:
        groupId: { type: ["string", "null"], minLength: 1, maxLength: 80 }
        assignmentDate: { type: string, pattern: "^\\d{4}-(0[1-9]|1[0-2])-(0[1-9]|[12]\\d|3[01])$" }
    PairingCommandInput:
      oneOf:
        - $ref: "#/components/schemas/AddPointsCommand"
        - $ref: "#/components/schemas/UndoCommand"
        - $ref: "#/components/schemas/ResetMonthCommand"
        - $ref: "#/components/schemas/SaveLapCommand"
        - $ref: "#/components/schemas/DeleteLapCommand"
        - $ref: "#/components/schemas/AddActivityCommand"
        - $ref: "#/components/schemas/DeleteActivityCommand"
        - $ref: "#/components/schemas/ClearActivitiesCommand"
        - $ref: "#/components/schemas/ToggleRewardCommand"
    AddPointsCommand:
      type: object
      additionalProperties: false
      required: [id, type, payload]
      properties:
        id: { type: string, format: uuid }
        type: { type: string, enum: [addPoints] }
        payload:
          type: object
          additionalProperties: false
          required: [groupId, amount, reason]
          properties:
            groupId: { type: string, minLength: 1, maxLength: 80 }
            amount: { type: number, minimum: -1000000, maximum: 1000000 }
            reason: { type: string, minLength: 1, maxLength: 500 }
    UndoCommand:
      type: object
      additionalProperties: false
      required: [id, type, payload]
      properties:
        id: { type: string, format: uuid }
        type: { type: string, enum: [undo] }
        payload:
          type: object
          additionalProperties: false
          required: [logId]
          properties:
            logId: { type: string, minLength: 1, maxLength: 80 }
    ResetMonthCommand:
      type: object
      additionalProperties: false
      required: [id, type, payload]
      properties:
        id: { type: string, format: uuid }
        type: { type: string, enum: [resetMonth] }
        payload:
          type: object
          additionalProperties: false
          required: [month]
          properties:
            month: { type: string, minLength: 1, maxLength: 100 }
    SaveLapCommand:
      type: object
      additionalProperties: false
      required: [id, type, payload]
      properties:
        id: { type: string, format: uuid }
        type: { type: string, enum: [saveLap] }
        payload:
          type: object
          additionalProperties: false
          required: [record]
          properties:
            record: { $ref: "#/components/schemas/LapRecord" }
    DeleteLapCommand:
      type: object
      additionalProperties: false
      required: [id, type, payload]
      properties:
        id: { type: string, format: uuid }
        type: { type: string, enum: [deleteLap] }
        payload:
          type: object
          additionalProperties: false
          required: [id]
          properties:
            id: { type: string, minLength: 1, maxLength: 80 }
    AddActivityCommand:
      type: object
      additionalProperties: false
      required: [id, type, payload]
      properties:
        id: { type: string, format: uuid }
        type: { type: string, enum: [addActivity] }
        payload:
          type: object
          additionalProperties: false
          required: [activity]
          properties:
            activity: { $ref: "#/components/schemas/Activity" }
    DeleteActivityCommand:
      type: object
      additionalProperties: false
      required: [id, type, payload]
      properties:
        id: { type: string, format: uuid }
        type: { type: string, enum: [deleteActivity] }
        payload:
          type: object
          additionalProperties: false
          required: [id]
          properties:
            id: { type: string, minLength: 1, maxLength: 80 }
    ClearActivitiesCommand:
      type: object
      additionalProperties: false
      required: [id, type, payload]
      properties:
        id: { type: string, format: uuid }
        type: { type: string, enum: [clearActivities] }
        payload:
          type: object
          additionalProperties: false
    ToggleRewardCommand:
      type: object
      additionalProperties: false
      required: [id, type, payload]
      properties:
        id: { type: string, format: uuid }
        type: { type: string, enum: [toggleReward] }
        payload:
          type: object
          additionalProperties: false
          required: [id]
          properties:
            id: { type: string, minLength: 1, maxLength: 80 }