# lib/api-zod/src/generated/types/clearActivitiesCommand.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ClearActivitiesCommandPayload } from './clearActivitiesCommandPayload';
import type { ClearActivitiesCommandType } from './clearActivitiesCommandType';

export interface ClearActivitiesCommand {
  id: string;
  type: ClearActivitiesCommandType;
  payload: ClearActivitiesCommandPayload;
}