# lib/api-zod/src/generated/types/deleteActivityCommand.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { DeleteActivityCommandPayload } from './deleteActivityCommandPayload';
import type { DeleteActivityCommandType } from './deleteActivityCommandType';

export interface DeleteActivityCommand {
  id: string;
  type: DeleteActivityCommandType;
  payload: DeleteActivityCommandPayload;
}