# lib/api-zod/src/generated/types/historyEntry.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface HistoryEntry {
  /**
     * @minLength 1
     * @maxLength 80
     */
  id: string;
  /**
     * @minLength 1
     * @maxLength 80
     */
  groupId: string;
  /**
     * @minLength 1
     * @maxLength 100
     */
  groupName: string;
  /**
     * @minimum -1000000
     * @maximum 1000000
     */
  amount: number;
  /**
     * @minLength 1
     * @maxLength 500
     */
  reason: string;
  /**
     * @minLength 1
     * @maxLength 100
     */
  timestamp: string;
}