# lib/api-zod/src/generated/types/lapRecord.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface LapRecord {
  /**
     * @minLength 1
     * @maxLength 80
     */
  id: string;
  /**
     * @minLength 1
     * @maxLength 100
     */
  runnerName: string;
  /**
     * @minLength 1
     * @maxLength 100
     */
  group: string;
  /**
     * @minimum 0
     * @maximum 999
     */
  minutes: number;
  /**
     * @minimum 0
     * @maximum 59
     */
  seconds: number;
  /**
     * @minimum 0
     * @maximum 99
     */
  ms: number;
  /**
     * @minLength 1
     * @maxLength 30
     */
  timeFormatted: string;
  /**
     * @minimum 0
     * @maximum 100000
     */
  totalSeconds: number;
  /**
     * @minLength 1
     * @maxLength 200
     */
  courseName: string;
  /**
     * @minLength 1
     * @maxLength 30
     */
  date: string;
  /**
     * @minLength 1
     * @maxLength 100
     */
  monthYear: string;
}